prevent removing selected items when ui is focussed

+++ BlueSwitch/BlueSwitch/Controls/Docking/RendererBase.cs
@@ -305,7 +305,7 @@ namespace BlueSwitch.Controls.Docking
 
         private void RendererBase_KeyUp(object sender, KeyEventArgs e)
         {
-            if (e.KeyCode == Keys.Delete)
+            if (e.KeyCode == Keys.Delete && !RenderingEngine.SelectionService.ItemWithFocusExists())
             {
                 RenderingEngine.SelectionService.RemoveSelected();
             }
